prep for new assets page
