+++ packages/web/components/complex/assets-page-v2.tsx
@@ -1,0 +1,6 @@
+import { observer } from "mobx-react-lite";
+import { FunctionComponent } from "react";
+
+export const AssetsPageV2: FunctionComponent = observer(() => {
+  return <main />;
+});
